@@ -11,7 +11,7 @@ keywords = ["archiving", "preservation", "identification"]
 include = ["pyproject.toml"]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 click = ">=7,<9"
 tqdm = "^4.36"
 dacite = "^1.0"
