@@ -1,28 +1,23 @@
-"""Identify files using
-`siegfried <https://github.com/richardlehane/siegfried>`
-"""
+"""Identify files using siegfried <https://github.com/richardlehane/siegfried>`."""
 import json
+import os
 import re
 import subprocess
-import os
 import warnings
-from logging import Logger
 from functools import partial
+from logging import Logger
+from multiprocessing import Pool
 from pathlib import Path
 from threading import Lock
-import PIL
-from PIL import Image
-from typing import Any, Tuple
-from typing import Dict
-from typing import List
-from multiprocessing import Pool
+from typing import Any
 
+import PIL
 from acamodels import Identification
+from PIL import Image
 
 from digiarch.core.ArchiveFileRel import ArchiveFile
 from digiarch.core.utils import natsort_path
 from digiarch.exceptions import IdentificationError
-
 
 warnings.filterwarnings("error", category=Image.DecompressionBombWarning)
 
@@ -38,9 +33,7 @@ RERUN_FORMATS = [
 SIZE_OF_KILO_BYTE = 1024
 
 
-def update_file_id(
-    path: Path, file_id: Identification, signature: Dict[str, str]
-) -> None:
+def update_file_id(path: Path, file_id: Identification, signature: dict[str, str]) -> None:
     file_id.puid = signature["puid"]
     file_id.signature = signature["signature"]
     if path.suffix.lower() != signature["extension"].lower():
@@ -51,7 +44,7 @@ def update_file_id(
 
 def custom_id(path: Path, file_id: Identification) -> Identification:
     sig_file = Path(__file__).parent / "custom_sigs.json"
-    signatures: List[Dict] = json.load(sig_file.open(encoding="utf-8"))
+    signatures: list[dict] = json.load(sig_file.open(encoding="utf-8"))
 
     with path.open("rb") as file_bytes:
         # BOF
@@ -72,10 +65,9 @@ def custom_id(path: Path, file_id: Identification) -> Identification:
                 if bof_pattern.search(bof) or eof_pattern.search(eof):
                     update_file_id(path, file_id, sig)
                     break
-            elif sig["operator"] == "AND":
-                if bof_pattern.search(bof) and eof_pattern.search(eof):
-                    update_file_id(path, file_id, sig)
-                    break
+            elif sig["operator"] == "AND" and bof_pattern.search(bof) and eof_pattern.search(eof):
+                update_file_id(path, file_id, sig)
+                break
         elif "bof" in sig:
             bof_pattern = re.compile(sig["bof"])
             if bof_pattern.search(bof):
@@ -89,30 +81,29 @@ def custom_id(path: Path, file_id: Identification) -> Identification:
     return file_id
 
 
-def sf_id(path: Path) -> Dict[Path, Identification]:
-    """Identify files using
-    `siegfried <https://github.com/richardlehane/siegfried>`_ and update
-    FileInfo with obtained PUID, signature name, and warning if applicable.
+def sf_id(path: Path) -> dict[Path, Identification]:
+    """Identify files using `siegfried <https://github.com/richardlehane/siegfried>`.
+
+    Alsio updates FileInfo with obtained PUID, signature name, and warning if applicable.
 
     Parameters
     ----------
     path : pathlib.Path
         Path in which to identify files.
 
-    Returns
+    Returns:
     -------
     Dict[Path, Identification]
         Dictionary containing file path and associated identification
         information obtained from siegfried's stdout.
 
-    Raises
+    Raises:
     ------
     IdentificationError
         If running siegfried or loading of the resulting JSON output fails,
         an IdentificationError is thrown.
     """
-
-    id_dict: Dict[Path, Identification] = {}
+    id_dict: dict[Path, Identification] = {}
 
     try:
         sf_proc = subprocess.run(
@@ -129,7 +120,7 @@ def sf_id(path: Path) -> Dict[Path, Identification]:
         raise IdentificationError(error)
 
     for file_result in id_result.get("files", []):
-        match: Dict[str, Any] = {}
+        match: dict[str, Any] = {}
         for id_match in file_result.get("matches"):
             if id_match.get("ns") == "pronom":
                 match = id_match
@@ -137,10 +128,7 @@ def sf_id(path: Path) -> Dict[Path, Identification]:
             file_identification: Identification
             file_path: Path = Path(file_result["filename"])
 
-            if match.get("id", "").lower() == "unknown":
-                puid = None
-            else:
-                puid = match.get("id")
+            puid = None if match.get("id", "").lower() == "unknown" else match.get("id")
 
             signature_and_version = None
             signature = match.get("format")
@@ -175,7 +163,8 @@ def sf_id(path: Path) -> Dict[Path, Identification]:
 
 
 def is_binary(file: ArchiveFile) -> bool:
-    """
+    """Check if a ArchiveFile is binary.
+
     Description:
     ----------------
     Checks if an ArchiveFile is a txt file or binary.
@@ -189,17 +178,12 @@ def is_binary(file: ArchiveFile) -> bool:
     word_markup_puid = "aca-fmt/2"
     bytes_of_file = file.read_bytes()
 
-    if b"\x00" in bytes_of_file:
+    if b"\x00" in bytes_of_file or (pdf_signature in bytes_of_file.hex()):
         return True
-    elif pdf_signature in bytes_of_file.hex():
-        return True
-    elif file.puid == word_markup_puid:
-        return True
-    else:
-        return False
+    return str(file.puid) == word_markup_puid
 
 
-def is_preservable(file: ArchiveFile, log: Logger) -> Tuple[bool, Any]:
+def is_preservable(file: ArchiveFile, log: Logger) -> tuple[bool, Any]:
     lock: Lock = Lock()  # Used by image_is_preservable
     image_format_codes = [
         "fmt/3",
@@ -241,15 +225,10 @@ def get_pixel_amount(file_path: Path) -> int:
 
 def check_if_preservable(file_path: Path) -> bool:
     pixel_amount = get_pixel_amount(file_path)
-    if pixel_amount < 20000:
-        return False
-    else:
-        return True
+    return pixel_amount > 20000
 
 
-def image_is_preservable(
-    file: ArchiveFile, lock: Lock, logger: Logger
-) -> bool:
+def image_is_preservable(file: ArchiveFile, lock: Lock, logger: Logger) -> bool:
 
     lock.acquire()
     result: bool = False
@@ -264,27 +243,17 @@ def image_is_preservable(
         print(f"PIL could not open the file: {file.relative_path}")
         result = True
     except Image.DecompressionBombWarning:
-        logger.warning(
-            "The file {} threw a decompresionbomb warning".format(
-                file.relative_path
-            )
-        )
+        logger.warning(f"The file {file.relative_path} threw a decompresionbomb warning")
         result = True
     except Image.DecompressionBombError:
-        logger.error(
-            "The file {} threw a decompresionbomb error".format(
-                file.relative_path
-            )
-        )
+        logger.error(f"The file {file.relative_path} threw a decompresionbomb error")
         result = True
     finally:
         lock.release()
         return result
 
 
-def update_file_info(
-    file_info: ArchiveFile, id_info: Dict[Path, Identification]
-) -> ArchiveFile:
+def update_file_info(file_info: ArchiveFile, id_info: dict[Path, Identification]) -> ArchiveFile:
     no_id: Identification = Identification(
         puid=None,
         signature=None,
@@ -305,7 +274,7 @@ def update_file_info(
     return file_info
 
 
-def identify(files: List[ArchiveFile], path: Path) -> List[ArchiveFile]:
+def identify(files: list[ArchiveFile], path: Path) -> list[ArchiveFile]:
     """Identify all files in a list, and return the updated list.
 
     Parameters
@@ -313,14 +282,13 @@ def identify(files: List[ArchiveFile], path: Path) -> List[ArchiveFile]:
     files : List[FileInfo]
         Files to identify.
 
-    Returns
+    Returns:
     -------
     List[FileInfo]
         Input files with updated Identification information.
 
     """
-
-    id_info: Dict[Path, Identification] = sf_id(path)
+    id_info: dict[Path, Identification] = sf_id(path)
     # functools.partial: Return a new partial object
     # which when called will behave like func called with the
     # positional arguments args and keyword arguments keywords.
@@ -331,6 +299,6 @@ def identify(files: List[ArchiveFile], path: Path) -> List[ArchiveFile]:
     # can be shorter than files.
     _update = partial(update_file_info, id_info=id_info)
     with Pool(5) as p:
-        updated_files: List[ArchiveFile] = p.map(_update, files, 10000)
+        updated_files: list[ArchiveFile] = p.map(_update, files, 10000)
 
     return natsort_path(updated_files)
