@@ -1,6 +1,4 @@
-"""Utilities for handling files, paths, etc.
-
-"""
+"""Utilities for handling files, paths, etc."""
 # -----------------------------------------------------------------------------
 # Imports
 # -----------------------------------------------------------------------------
@@ -8,26 +6,21 @@ import os
 import shutil
 from datetime import datetime
 from pathlib import Path
-from typing import List
 
-
-from digiarch.core.ArchiveFileRel import ArchiveFile
-from digiarch.core.utils import natsort_path
-from digiarch.core.utils import size_fmt
-from digiarch.exceptions import FileCollectionError
-from digiarch.models import FileData
-from digiarch.models import Metadata
 from tqdm import tqdm
 
+from digiarch.core.ArchiveFileRel import ArchiveFile
+from digiarch.core.utils import natsort_path, size_fmt
+from digiarch.exceptions import FileCollectionError
+from digiarch.models import FileData, Metadata
 
 # -----------------------------------------------------------------------------
 # Function Definitions
 # -----------------------------------------------------------------------------
 
 
-async def explore_dir(file_data: FileData) -> List[str]:
-    """Finds files and empty directories in the given path,
-    and writes them to a file database.
+async def explore_dir(file_data: FileData) -> list[str]:
+    """Finds files and empty directories in the given path, and writes them to a file database.
 
     Parameters
     ----------
@@ -35,28 +28,24 @@ async def explore_dir(file_data: FileData) -> List[str]:
         File data model including main directory, data directory,
         and file database.
 
-    Returns
+    Returns:
     ----------
     List[str]
         List of warnings encountered while parsing directories.
 
     """
     # Type declarations
-    dir_info: List[ArchiveFile] = []
-    empty_subs: List[Path] = []
-    multiple_files: List[Path] = []
-    warnings: List[str] = []
+    dir_info: list[ArchiveFile] = []
+    empty_subs: list[Path] = []
+    multiple_files: list[Path] = []
+    warnings: list[str] = []
     total_size: int = 0
     file_count: int = 0
     main_dir: Path = file_data.main_dir
     data_dir: Path = file_data.data_dir
-    metadata = Metadata(
-        last_run=datetime.now(), processed_dir=file_data.main_dir
-    )
+    metadata = Metadata(last_run=datetime.now(), processed_dir=file_data.main_dir)  # noqa: DTZ005
 
-    if not [
-        child for child in main_dir.iterdir() if child.name != data_dir.name
-    ]:
+    if not [child for child in main_dir.iterdir() if child.name != data_dir.name]:
         # Path is empty, remove main directory and raise
         shutil.rmtree(data_dir)
         raise FileCollectionError(f"{main_dir} is empty! No files collected.")
@@ -65,7 +54,9 @@ async def explore_dir(file_data: FileData) -> List[str]:
     # tqdm is used to show progress of os.walk
 
     for root, dirs, files in tqdm(
-        os.walk(main_dir, topdown=True), unit=" folders", desc="Processed"
+        os.walk(main_dir, topdown=True),
+        unit=" folders",
+        desc="Processed",
     ):
 
         if data_dir.name in dirs:
@@ -79,9 +70,7 @@ async def explore_dir(file_data: FileData) -> List[str]:
         for file in files:
             try:
                 cur_path = Path(root, file)
-                cur_rel_path = cur_path.relative_to(
-                    Path(os.environ["ROOTPATH"])
-                )
+                cur_rel_path = cur_path.relative_to(Path(os.environ["ROOTPATH"]))
                 dir_info.append(ArchiveFile(relative_path=cur_rel_path))
             except Exception as e:
                 raise FileCollectionError(e)
