@@ -1,18 +1,18 @@
-"""Shared testing fixtures.
-
-"""
+"""Shared testing fixtures."""
 # -----------------------------------------------------------------------------
 # Imports
 # -----------------------------------------------------------------------------
+import os
+from logging import Logger
+from pathlib import Path
 from threading import Lock
-from typing import Tuple, Union
+from typing import Union
+
+import png
+import pytest
+
 from digiarch.cli import setup_logger
 from digiarch.core.ArchiveFileRel import ArchiveFile
-from pathlib import Path
-from logging import Logger
-import pytest
-import os
-import png
 
 # -----------------------------------------------------------------------------
 # Function Definitions
@@ -21,7 +21,7 @@ import png
 pytestmark = pytest.mark.asyncio
 
 
-@pytest.fixture
+@pytest.fixture()
 def temp_dir(tmpdir_factory):
     t_dir: str = tmpdir_factory.mktemp("temp_dir")
     temp_dir: Path = Path(t_dir, "AARS.TEST")
@@ -29,14 +29,14 @@ def temp_dir(tmpdir_factory):
     return temp_dir
 
 
-@pytest.fixture
+@pytest.fixture()
 def main_dir(temp_dir):
     main_dir: Path = temp_dir / "_metadata"
     main_dir.mkdir(exist_ok=True)
     return main_dir
 
 
-@pytest.fixture
+@pytest.fixture()
 def data_file(main_dir):
     data_dir: Path = main_dir / ".data"
     data_dir.mkdir(exist_ok=True)
@@ -44,43 +44,43 @@ def data_file(main_dir):
     return data_file
 
 
-@pytest.fixture
+@pytest.fixture()
 def test_data_dir():
     return Path(__file__).parent.parent / "tests" / "_data" / "AARS.test_data"
 
 
-@pytest.fixture
+@pytest.fixture()
 def docx_info(test_data_dir):
     docx_file: Path = test_data_dir / "docx_test.docx"
     return docx_file
 
 
-@pytest.fixture
+@pytest.fixture()
 def python_wiki(test_data_dir):
     python_wiki: Path = test_data_dir / "Python_wiki.docx"
     return python_wiki
 
 
-@pytest.fixture
+@pytest.fixture()
 def xls_info(test_data_dir):
     xls_file: Path = test_data_dir / "xls_test.xls"
     return xls_file
 
 
-@pytest.fixture
+@pytest.fixture()
 def adx_info(test_data_dir):
     adx_file: Path = test_data_dir / "adx_test.adx"
     return adx_file
 
 
-@pytest.fixture
+@pytest.fixture()
 def file_data(temp_dir):
     from digiarch.models import FileData
 
     return FileData(main_dir=temp_dir, files=[])
 
 
-@pytest.fixture
+@pytest.fixture()
 def non_binary_file():
     os.environ["ROOTPATH"] = str(Path.cwd())
     text_file_relative_path = Path("text_file.txt")
@@ -92,7 +92,7 @@ def non_binary_file():
     text_file_path.unlink()
 
 
-@pytest.fixture
+@pytest.fixture()
 def binary_file():
     os.environ["ROOTPATH"] = str(Path.cwd())
     image_file_relative_path = Path("image_file.png")
@@ -103,9 +103,9 @@ def binary_file():
     height = 255
     img = []
     for y in range(height):
-        row: Union[Tuple[int, int, int], Tuple] = ()
+        row: Union[tuple[int, int, int], tuple] = ()
         for x in range(width):
-            row = row + (x, max(0, 255 - x - y), y)
+            row = (*row, x, max(0, 255 - x - y), y)
         img.append(row)
     with open(image_file_path, "wb") as f:
         w = png.Writer(width, height, greyscale=False)
@@ -119,7 +119,7 @@ def binary_file():
     image_file_path.unlink()
 
 
-@pytest.fixture
+@pytest.fixture()
 def small_binary_file():
     os.environ["ROOTPATH"] = str(Path.cwd())
     image_file_relative_path = Path("image_file.png")
@@ -130,9 +130,9 @@ def small_binary_file():
     height = 50
     img = []
     for y in range(height):
-        row: Union[Tuple[int, int, int], Tuple] = ()
+        row: Union[tuple[int, int, int], tuple] = ()
         for x in range(width):
-            row = row + (x, max(0, 255 - x - y), y)
+            row = (*row, x, max(0, 255 - x - y), y)
         img.append(row)
     with open(image_file_path, "wb") as f:
         w = png.Writer(width, height, greyscale=False)
@@ -146,7 +146,7 @@ def small_binary_file():
     image_file_path.unlink()
 
 
-@pytest.fixture
+@pytest.fixture()
 def very_small_binary_file():
     os.environ["ROOTPATH"] = str(Path.cwd())
     file_relative_path = Path("very_small_file.claus")
@@ -166,16 +166,15 @@ def very_small_binary_file():
     file_path.unlink()
 
 
-@pytest.fixture
+@pytest.fixture()
 def python_wiki_binary_file(python_wiki):
 
-    larger_binary_file = ArchiveFile(
+    return ArchiveFile(
         relative_path=python_wiki,
         puid="fmt/fmt/412",
         signature="Docx file",
         is_binary=True,
     )
-    return larger_binary_file
 
 
 @pytest.fixture(scope="session")
