@@ -2,8 +2,7 @@
 # Imports
 # -----------------------------------------------------------------------------
 from digiarch.core.ArchiveFileRel import ArchiveFile
-from digiarch.core.utils import natsort_path
-from digiarch.core.utils import size_fmt
+from digiarch.core.utils import natsort_path, size_fmt
 
 # -----------------------------------------------------------------------------
 # Tests
